requirements.txt
pyinfra==2.7
python-dotenv==1.0.0